Updated mdata/igrf.d: 1980-2020 with 2020 being an extrapolation!
Mike Hapgood's public version on http://www.jsoc.rl.ac.uk/pub/ovt/.

NOTE: The data for year 2020 are extrapolations!
Hence the data should be updated eventually with actual
This should NOT be confused with the additional column with first
derivatives used for extrapolation after the last year of data.